src/notification/discord_notifier.py
import discord

class DiscordNotifier:
    def __init__(self, token):
        self.client = discord.Client()
        self.token = token

    async def send_notification(self, channel_id, message):
        # Placeholder function for sending Discord notifications
        pass
