requirements.txt
requests==2.26.0
beautifulsoup4==4.10.0
pandas==1.3.3
nltk==3.6.3
spacy==3.1.3
apscheduler==3.8.1
jinja2==3.0.2
discord.py==1.7.3
pyyaml==6.0
pytest==6.2.5
